requirements.txt
aniso8601==8.0.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-Ext==0.1
Flask-RESTful==0.3.8
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
pyodbc==4.0.30
python-dotenv==0.14.0
pytz==2020.1
six==1.15.0
waitress==1.4.4
Werkzeug==1.0.1
WTForms==2.3.3
tinydb==4.4.0
PyJWT
pytest==6.2.3
mock
requests